................
  r28912 | res2002 | 2008-03-07 19:36:24 +0100 (Fri, 07 Mar 2008) | 3 lines
  
  Initialized merge tracking via "svnmerge" with revisions "1-28908" from 
  https://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk
................
  r28918 | res2002 | 2008-03-09 16:39:15 +0100 (Sun, 09 Mar 2008) | 2 lines
  
  res added some helpers to create matrices to common projections.
................
  r28920 | res2002 | 2008-03-09 17:01:42 +0100 (Sun, 09 Mar 2008) | 2 lines
  
  res started to make iCamera projection-agnostic.
................
  r28921 | res2002 | 2008-03-09 17:08:03 +0100 (Sun, 09 Mar 2008) | 2 lines
  
  res added a to get an OpenGL matrix from a CS::Math::Matrix4.
................
  r28922 | res2002 | 2008-03-09 17:15:25 +0100 (Sun, 09 Mar 2008) | 2 lines
  
  res started to add explicit projection matrix support to iGraphics3D.
................
  r28923 | res2002 | 2008-03-09 18:05:34 +0100 (Sun, 09 Mar 2008) | 2 lines
  
  Use actually correct projection matrices
................
  r28925 | res2002 | 2008-03-09 18:36:08 +0100 (Sun, 09 Mar 2008) | 30 lines
  
  Merged revisions 28909,28911,28913,28919,28924 via svnmerge from 
  https://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk
  
  ........
    r28909 | sunshine | 2008-03-07 19:14:26 +0100 (Fri, 07 Mar 2008) | 2 lines
    
    Minor documentation and status reporting corrections.
  ........
    r28911 | sunshine | 2008-03-07 19:24:08 +0100 (Fri, 07 Mar 2008) | 6 lines
    
    Eric Sunshine changed CS/docs/support/prepare_online_publish.sh so that it
    no longer converts published HTML files to PHP. This conversion can be
    avoided since we no longer utilize PHP in the online documentation. (It
    had been used by the now-retired annotation facility.) Furthermore,
    vending HTML rather than PHP reduces server load.
  ........
    r28913 | res2002 | 2008-03-08 19:12:13 +0100 (Sat, 08 Mar 2008) | 3 lines
    
    res fixed a pattern in CS/docs/support/prepare_online_publish.sh that
      was still using a .php instead of a .html file name.
  ........
    r28919 | thebolt00 | 2008-03-09 16:42:16 +0100 (Sun, 09 Mar 2008) | 2 lines
    
    - Marten fixed bug #461, terrain2 setting wrong parameter.
  ........
    r28924 | res2002 | 2008-03-09 18:22:04 +0100 (Sun, 09 Mar 2008) | 2 lines
    
    res actually implemented CS::Math::Matrix4::Description().
  ........
................
  r28926 | res2002 | 2008-03-09 18:37:40 +0100 (Sun, 09 Mar 2008) | 2 lines
  
  res implemented explicit projection matrix support for gl3d.
................
  r28927 | res2002 | 2008-03-09 18:39:57 +0100 (Sun, 09 Mar 2008) | 3 lines
  
  res changed rendering by the engine to set an explicit projection
    matrix.
................
  r28928 | res2002 | 2008-03-09 18:54:46 +0100 (Sun, 09 Mar 2008) | 2 lines
  
  res implemented explicit projection matrix support for null3d.
................
  r28929 | res2002 | 2008-03-09 19:03:22 +0100 (Sun, 09 Mar 2008) | 2 lines
  
  Some fixes
................
  r28931 | res2002 | 2008-03-09 19:23:00 +0100 (Sun, 09 Mar 2008) | 10 lines
  
  Merged revisions 28930 via svnmerge from 
  https://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk
  
  ........
    r28930 | res2002 | 2008-03-09 19:14:52 +0100 (Sun, 09 Mar 2008) | 4 lines
    
    res corrected the creation of a Matrix4 from a transform.
    res added multiplication of a Matrix4 with a csVector4.
  ........
................
  r28932 | res2002 | 2008-03-09 20:42:14 +0100 (Sun, 09 Mar 2008) | 2 lines
  
  res implemented explicit projection matrix support for soft3d.
................
  r28933 | res2002 | 2008-03-09 21:40:15 +0100 (Sun, 09 Mar 2008) | 3 lines
  
  res added clip plane providing by the camera, depending on the
    actual projection. (Unused and incomplete tho.)
................
  r28938 | res2002 | 2008-03-09 22:31:14 +0100 (Sun, 09 Mar 2008) | 3 lines
  
  res added methods to the projection/custom matrix interfaces to
    quickly obtain the iCamera interface.
................
  r28940 | res2002 | 2008-03-10 13:54:27 +0100 (Mon, 10 Mar 2008) | 13 lines
  
  Merged revisions 28934,28939 via svnmerge from 
  https://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk
  
  ........
    r28934 | crystal-manage | 2008-03-09 21:41:56 +0100 (Sun, 09 Mar 2008) | 2 lines
    
    Retired highly outdated docs/todo_newrenederer.txt.
  ........
    r28939 | res2002 | 2008-03-10 13:42:14 +0100 (Mon, 10 Mar 2008) | 2 lines
    
    res added inverting to CS::Math::Matrix4.
  ........
................
  r28945 | res2002 | 2008-03-11 12:10:48 +0100 (Tue, 11 Mar 2008) | 12 lines
  
  res made the following changes:
    - Changed the iCamera interface so that getting the projection
      matrix doesn't require the screen dimensions any more.
    - iProjectionCamera now takes FOV, perspective center etc. in
      normalized screen space, not pixels. (Note: old iCamera still
      takes pixel coords!)
    - Added visible volume determination for custom matrix camera.
    - Changed how RenderViewClipper works: it now takes the visible
      volume from the camera instead of the frustum computed by the
      render view.
................
  r28946 | res2002 | 2008-03-11 12:12:05 +0100 (Tue, 11 Mar 2008) | 12 lines
  
  res made the following changes:
    - Changed the iCamera interface so that getting the projection
      matrix doesn't require the screen dimensions any more.
    - iProjectionCamera now takes FOV, perspective center etc. in
      normalized screen space, not pixels. (Note: old iCamera still
      takes pixel coords!)
    - Added visible volume determination for custom matrix camera.
    - Changed how RenderViewClipper works: it now takes the visible
      volume from the camera instead of the frustum computed by the
      render view.
................


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/feature/rendermanager@28965 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csgeom/projections.h
@@ -55,7 +55,7 @@ namespace CS
       /**
        * Create a perspective projection as used by CS.
        */
-      static Matrix4 CSPerspective (int viewWidth, int viewHeight, 
+      static Matrix4 CSPerspective (float viewWidth, float viewHeight, 
 				    float shiftX, float shiftY,
 				    float invAspect)
       {
@@ -83,3 +83,4 @@ namespace CS
 /** @} */
 
 #endif // __CS_CSGEOM_PROJECTIONS_H__
+
